+++ src/app/app.module.ts
@@ -1,9 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { HttpClientModule } from '@angular/common/http';
 
-import { FeaturesModule } from './features/features.module';
+// import { FeaturesModule } from './features/features.module';
 import { AppRoutingModule } from './/app-routing.module';
 
 import { AppComponent } from './app.component';
@@ -11,14 +9,7 @@ import { CoreModule } from '@core/core.module';
 
 @NgModule({
 	declarations: [AppComponent],
-	imports: [
-		BrowserModule,
-		FeaturesModule,
-		CoreModule,
-		AppRoutingModule,
-		BrowserAnimationsModule,
-		HttpClientModule
-	],
+	imports: [BrowserModule, AppRoutingModule, CoreModule],
 	providers: [],
 	bootstrap: [AppComponent]
 })
