+++ src/app/core/modal/modal-views/create-issue/create-issue.component.sass
