+++ src/app/core/modal/modal-views/worker-view/worker-view.component.sass
