+++ src/app/core/modal/modal.component.sass
