+++ src/app/core/modal/modal.component.ts
@@ -1,24 +1,35 @@
-import { Component, ViewChild, Input, ElementRef, Output, EventEmitter, OnInit, ComponentFactoryResolver, Type, Renderer2, AfterContentInit } from '@angular/core';
+import {
+	Component,
+	ViewChild,
+	Input,
+	ElementRef,
+	Output,
+	EventEmitter,
+	OnInit,
+	ComponentFactoryResolver,
+	Type,
+	Renderer2,
+	AfterContentInit
+} from '@angular/core';
 import { ModalDirective } from './modal.directive';
 import { CreateIssueComponent } from './modal-views/create-issue/create-issue.component';
-import { ModalType } from '../../core/enums/modal-type';
+import { ModalType } from '../enums/modal-type';
 import { UpdateIssueComponent } from './modal-views/update-issue/update-issue.component';
-import { Issue } from '../../core/models/issue';
+import { Issue } from '../models/issue';
 import { WorkerViewComponent } from './modal-views/worker-view/worker-view.component';
 
 @Component({
-  selector: 'app-modal',
-  template: `
+	selector: 'app-modal',
+	template: `
 		<div #container class="overlay">
 			<div class="modal">
 				<ng-template modal-host></ng-template>
 			</div>
 		</div>
-  `,
-  styleUrls: ['./modal.component.sass']
+	`,
+	styleUrls: ['./modal.component.sass']
 })
 export class ModalComponent implements OnInit, AfterContentInit {
-
 	@Input() innerComponentType: ModalType;
 	@Input() issue: Issue;
 	@ViewChild(ModalDirective, { static: true }) modalHost: ModalDirective;
@@ -37,34 +48,51 @@ export class ModalComponent implements OnInit, AfterContentInit {
 	}
 
 	ngAfterContentInit(): void {
-		if (this.innerComponentType === ModalType.UPDATE || this.innerComponentType === ModalType.WORKER) {
+		if (
+			this.innerComponentType === ModalType.UPDATE ||
+			this.innerComponentType === ModalType.WORKER
+		) {
 			this.innerComponent.issue = this.issue;
 		}
 	}
-	
+
 	public closeModal(canceled: boolean) {
 		this.container.nativeElement.style.animationName = 'fadeOut';
-		this.container.nativeElement.addEventListener('animationend', () => {
-					this.afterClose.emit(this.innerComponent.onClose(canceled));
-			}, false);
+		this.container.nativeElement.addEventListener(
+			'animationend',
+			() => {
+				this.afterClose.emit(this.innerComponent.onClose(canceled));
+			},
+			false
+		);
 	}
 
 	loadComponent(): void {
 		let viewContainerRef = this.modalHost.viewContainerRef;
 		viewContainerRef.clear();
-		const factory = this.componentFactoryResolver.resolveComponentFactory(this.modalTypeResolver(this.innerComponentType));
+		const factory = this.componentFactoryResolver.resolveComponentFactory(
+			this.modalTypeResolver(this.innerComponentType)
+		);
 		let insertRes = this.modalHost.viewContainerRef.createComponent(factory);
 		this.innerComponent = insertRes.instance;
-		this.renderer.listen(this.innerComponent.closeButton.nativeElement, 'click', (event) => {
-			this.closeModal(true);
-		});
-		this.renderer.listen(this.innerComponent.submitButton.nativeElement, 'click', (event) => {
-			this.closeModal(false);
-		})
+		this.renderer.listen(
+			this.innerComponent.closeButton.nativeElement,
+			'click',
+			event => {
+				this.closeModal(true);
+			}
+		);
+		this.renderer.listen(
+			this.innerComponent.submitButton.nativeElement,
+			'click',
+			event => {
+				this.closeModal(false);
+			}
+		);
 	}
 
 	private modalTypeResolver(type: ModalType): Type<{}> {
-		switch(type) {
+		switch (type) {
 			case ModalType.CREATE: {
 				return CreateIssueComponent;
 			}
@@ -79,5 +107,4 @@ export class ModalComponent implements OnInit, AfterContentInit {
 			}
 		}
 	}
-
 }
