+++ src/app/core/services/index.ts
@@ -2,3 +2,4 @@ export * from './issue.service';
 export * from './map.service';
 export * from './user.service';
 export * from './translation.service';
+export * from './master.service';
