+++ src/app/features/authorization/authorization.component.html
@@ -1,0 +1,3 @@
+<p>
+  authorization works!
+</p>
