+++ src/app/features/authorization/authorization.component.ts
@@ -1,0 +1,19 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-authorization',
+  template: `
+   <p>
+  		authorization Works!
+   </p>
+  `,
+  styleUrls: ['./authorization.component.sass']
+})
+export class AuthorizationComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
