+++ src/app/features/authorization/registration/registration.component.ts
