+++ src/app/features/dashboard/containers/index.ts
@@ -1,0 +1,6 @@
+export * from './automatisation/automatisation.component';
+export * from './map/map.component';
+export * from './overview/overview.component';
+export * from './profile/profile.component';
+export * from './statistics/statistics.component';
+export * from './tasks/tasks.component';
