+++ src/app/features/report/report-routing.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { ReportComponent } from './report.component';
+
+const routes: Routes = [
+	{
+		path: '',
+		component: ReportComponent
+	}
+];
+
+@NgModule({
+	imports: [RouterModule.forChild(routes)],
+	exports: [RouterModule]
+})
+export class ReportRoutingModule {}
