+++ src/app/features/report/report.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { ReportRoutingModule } from './report-routing.module';
+
+import { ReportComponent } from './report.component';
+
+@NgModule({
+	imports: [CommonModule, ReportRoutingModule],
+	declarations: [ReportComponent]
+})
+export class ReportModule {}
