+++ src/app/features/static/static.component.html
@@ -1,0 +1,3 @@
+<p>
+  static works!
+</p>
