+++ src/app/features/static/static.component.ts
@@ -1,0 +1,19 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-static',
+  template: `
+   <p>
+  		static Works!
+   </p>
+  `,
+  styleUrls: ['./static.component.sass']
+})
+export class StaticComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
